576: Improving test com.forgerock.securebanking.uk.gateway.jwks.cache.caffeine.CaffeineCacheTest.shouldEvictEntriesBasedOnTime
The test works when run locally, but we have noticed failures when run via a GitHub action.
The failures appear to be due to the test taking longer than expected to add the items to the cache, which means that some have been evicted before expected.

Fixing by increasing the expiryDuration and doing a sanity test that the cache has items (rather than checking for an exact number)

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/576