257: Create FapiTransactionIdFilter and use in test route
Create a filter as suggested by Guillaume Sauthier in the following
issue: https://bugster.forgerock.org/jira/browse/OPENIG-6538

The aim is to set the transactionIdContext to be the value of the
x-fapi-interaction-id. This _should_ ensure that the
value of the x-fapi-interaction-id assigned by IG is passed to AM & IDM
as the x-forgerock-transactionid header and will make logs traceable
across the system.

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/257