257: Use filter chain at start of routes
Use a filter chain at the start of the routes to ensure the
x-fapi-interaction-id is set and that the x-forgerock-transactionid is
set to the same value. These values need to be set at the start of the
route so that any scripts called, or any calls from filter to the
FRPlatform carry the x-fapi-interaction-id value.

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/257