656: Support PUT in CreateApiClient filter
PUT is used to update an existing DCR.

The CreateApiClient filter needs to update the IDM apiClient object for the DCR being updated. This logic is done on the response path (same as for POST), so that we only do the update if AM accepted the DCR update.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/656