678: Disabling message capture for IG route 99 (#254)
* 694: Disabling message capture for IG route 99

Route 99 is the "pass through" route for AM for access which does not have a more specific route. This includes accessing the XUI.

Adding new handler SBATReverseProxyHandlerIdentityPlatformNoCapture and using it in route 99 in order to stop capturing these messages in the logs. Access to the XUI includes all calls to fetch html/css/js resources and if capturing enables this floods the logs.

Also fixing a bug with the SBATReverseProxyHandlerIdentityPlatform and SBATReverseProxyHandlerRs handlers which were logging captures twice, this is due to them have capturing enabled and the underlying ReverseProxyHandler also having capturing enabled.

https://github.com/SecureApiGateway/SecureApiGateway/issues/699