729: Adding FetchTrustedDirectoryFilter
This filter will add the TrustedDirectory for this ApiClient to the attributes context. NOTE: FetchTrustedDirectoryFilter should be installed in a filter chain after the FetchApiClientFilter filter, as the ApiClient data is required to determine the TrustedDirectory.

Added filter chain: FetchApiClientAndTrustedDirectoryDataChain which can be used to fetch the ApiClient and then the TrustedDirectory. Updating existing usages of the FetchApiClientFilter in routes to use the new chain.

https://github.com/SecureApiGateway/SecureApiGateway/issues/729