773: Use sofware_id rather than software_client_id in ApiClient
The ApiClient should store the software_id rather than the software
client id. This is inkeeping with the OB SSA. It is unclear what
the software_client_id is for in the OB SSA - it seems to be the oauth2
clientId for the Software Statement as held by the Open Banking
Directory hence it is meanlingless for our cert issuance.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/733