773: process registration should not rely on certificate (#265)
* 691: Use TrustedDirectoriesService to get directory config

Issue: https://github.com/secureapigateway/secureapigateway/issues/691

* 691: Pass the TrustedDirectoriesService to ProcessRegistration

Issue: https://github.com/secureapigateway/secureapigateway/issues/691

* 691: Use 'software_id' from SAPI-G SSA for unique id

The software client id seems to be something different.

Issue: https://github.com/secureapigateway/secureapigateway/issues/691

* 691: Fixing scope validation in ProcessRegistration

We should not check the scopes in the registration request against the
roles in the eIDAS cert, as that is brittle. Rather we should check the
scopes against the roles found in the SSA

Issue: https://github.com/secureapigateway/secureapigateway/issues/733

* 773: Check request scopes against ssa roles

The script used to check the requested scopes against the roles in the
eidas certificate. This is brittle. We should check the requested scopes
against the roles in the SSA

Issue: https://github.com/secureapigateway/secureapigateway/issues/733

* 773: Fixes to ProcessRegistration and improvements

Clearer naming of the getJwksUri method on TrustedDirectory

Issue: https://github.com/secureapigateway/secureapigateway/issues/733

* 773: Use sofware_id rather than software_client_id in ApiClient

The ApiClient should store the software_id rather than the software
client id. This is inkeeping with the OB SSA. It is unclear what
the software_client_id is for in the OB SSA - it seems to be the oauth2
clientId for the Software Statement as held by the Open Banking
Directory hence it is meanlingless for our cert issuance.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/733