733: Fix AM fails to reach OB jwks_uri
We need AM to be able to reach the OB jwks_uri. There was some 'odd'
code in ProcessRegistration.groovy that re-wrote the jwks_uri, and we
din't think it was necessary. However it is necessary, as it means that
AM reaches out to a proxy route in IG that then goes and retrieves the
jwks_uri. This is because AM doesn't have the OB certs in it's trust
store.

I have put that code back
Issue: https://github.com/secureapigateway/secureapigateway/issues/733