740: Fixing signing algorithm validation bug in FAPIAdvancedDCRValidationFilter
The token_endpoint_auth_signing_alg field is optional in the DCR request. Changing validation logic for signing fields to skip fields if they are not present in the request.

The job of the FAPIAdvancedDCRValidationFilter is to reject requests that are not FAPI compliant, it is not expected to implement the DCR logic. Therefore, checking if fields are set correctly as per the OAuth2 DCR spec is a job for the filter that implements this logic.

https://github.com/SecureApiGateway/SecureApiGateway/issues/740