698: Removing cnf_key param from call to AM token endpoint
The cnf_key param is redundant, AM is configured to read the client mTLS cert from the ssl-client-cert header. Therefore, AM is able to compute the hash itself.

Removing the filters from the token endpoint chain which compute and set the hash param, removing AddCnfKey.groovy script as it was only used in this chain.

https://github.com/SecureApiGateway/SecureApiGateway/issues/698