765: Allow the Test Trusted Directory to be disabled by an env var
IG_TEST_DIRECTORY_ENABLED controls whether the Test Trusted Directory is enabled.

Note: this disables use of the directory in DCR (and OB API calls), as well as disabling the config from the TrustedDirectoryService. The jwkms related routes will still exist and be callable, but the certs produced will be unusable.

https://github.com/SecureApiGateway/SecureApiGateway/issues/765