733: New filter puts RegistrationRequest on context attribute
This PR addresses the PR comments in
https://github.com/SecureApiGateway/securebanking-openbanking-uk-gateway/pull/268

The DCR route uses a lot of filters that check fields in both the
registration request jwt and the software statement assertion. They all
do the work of turning the b64 encoded representation of the software
statement into a ClaimsSet and then using the TrustedDirectory for the
issuer to obtain the necessary fields to validate the SSA or get the
software statements jwks or jwks_uri.

This new filter, the RegistrationRequestEntityValidatorFilter will do
all that work up front at the start of the route and make a
RegistrationRequest object available as a context attribute for other
routes to use.

Issue: https://github.com/secureapigateway/secureapigateway/issues/733