793: Adding filter to check that the end user is authorised to access the consent request
Utilising IdTokenValidationFilter to validate the consent_request JWT, this filter verifies the signature and validates the "exp", "iat", "iss" and "aud claims".

The ConsentRequestAccessAuthorisationFilter uses the SsoTokenContext and JwtValidationContext set by the SingleSignOnFilter and IdTokenValidationFilter respectively. Using these contexts the filter verifies that the userId of the SSOToken matches that of the consent_request JWT signed by AM.

https://github.com/SecureApiGateway/SecureApiGateway/issues/793