805: Adding route 83-rcs-api.json which protects the RCS API
The new route does authentication and authorisation checks on calls made to the RCS API.

A new filter has been added to the route: RcsApiConsentRequestJwtResolver which resolves the consent_request JWT for the ConsentRequestJwtValidationFilter to validate it.

To do the authorisation checks we are validating the consent_request JWT. This means the RCS backend can trust that the JWT has had its signature verified.

https://github.com/SecureApiGateway/SecureApiGateway/issues/793