831: Adding CertificateResolver and HeaderCertificateResolver
HeaderCertificateResolver resolves the client's mTLS certificate from a HTTP request header. The resolver will decode the certificate and return an X509Certificate object, or it will throw an exception.

Deprecating the CertificateFromHeaderSupplier which does a similar job but only returns the raw header string, leaving it up to the caller to parse the cert.

The new HeaderCertificateResolver can be used to reduce code duplication involved in parsing the cert in multiple places.

https://github.com/SecureApiGateway/SecureApiGateway/issues/831