831: Initial implementation of TokenEndpointTransportCertValidationFilter
TokenEndpointTransportCertValidationFilter is a specialised form of TransportCertValidationFilter, this filter works on the response path and fetches the data needed to do the validation by calling services directly (rather than accessing values from the context). This is required because for the token endpoint route, we do not know that the client has been authenticated until we get the response back from AM, therefore it is safest do the validation on the response path.

https://github.com/SecureApiGateway/SecureApiGateway/issues/831