691: Fixing ProcessDetachedSig header claim: http://openbanking.org.uk/iss validation
http://openbanking.org.uk/iss header claim is validated to check it matches "{{org-id}}/{{software-statement-id}}"

To do this we need to get the ApiClient data, as the ProcessDetachedSig filter runs after FetchApiClientFilter then this data is available in the attributes context.

Fixing bug in IDM query which was not retrieiving the fields within the apiClientOrg object.

Fixing bug in ProcessDetachedSig which caused validation errors to be ignored as the isJwsSignatureValid was returning a Promise wrapping a boolean instead of a boolean.

https://github.com/SecureApiGateway/SecureApiGateway/issues/691