684: Creating SapiLogAttachedExceptionFilter alias for LogAttachedExceptionFilter (#326)
* 684: Creating SapiLogAttachedExceptionFilter alias for LogAttachedExceptionFilter

LogAttachedExceptionFilter does not have a heaplet, in order to control where it gets installed in a filter chain then a heaplet is required to be able to create it in config.

Installing the filter after FapiTransactionIdFilter. This means that when it catches an unhandled exception, it will log the x-fapi-interaction-id. Without this change a randomly generated transactionId is logged which makes it impossible to find when searching by x-fapi-interaction-id.

The existing LogAttachedExceptionFilter that IG installs by default will handle any unexpected exceptions that occur between the 2 filters.

https://github.com/SecureApiGateway/SecureApiGateway/issues/684