865: Refactor testing of RegistrationRequest and SoftwareStatement
TestFactories are now used to generate SoftwareStatements and
RegistrationRequests for use in tests. Previously it was a bit of a mix
of mocks etc, and a bit ad-hock, which meant changes to the required
claims for either a software statement or a registration request meant
fixing tests in a number of places which was time consuming. There is
now a framework that will create valid RegistrationReqeusts or
SoftwareSatements with the ability to override claims within them.

Issue: https://github.com/secureapigateway/secureapigateway/issues/865