865: Store SSA roles in the ApiClient object
Roles are needed for access decisions, and the roles in the SSA
ultimately limit the roles that a client can request and the apis that a
client can access. Each API request needs to know which roles were in
the Software Statement and limit access accordingly.

Both roles and the jwks are authentication and authorization information
and should be field in the ApiClient. This saves each route having to
parse the SSA and find the information from that jwt.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/865