873: Fixing NPE in GetResourceOwnerIdFromConsent amount logging
Logger line which is logging the amount for funds-confirmation flows thrown an NPE when called for VRP payment.

Fixing the logging to work with all payment types.

Fixing error response blocks to return a Promise when inside a thenAsync block.

Fixing catch Exception block to return a response (it was building one but not returning it).

https://github.com/SecureApiGateway/SecureApiGateway/issues/873