874: Adding X-Forwarded-Prefix header to RS RP requests
X-Forwarded-Prefix is used to contain a baseUri that exists on the gateway/ingress that has been removed when calling the upstream service (RS).

This fixes an issue with the Payments API links being generated by the RS being incorrect (missing the /rs path prefix). Spring HATEOAS impl uses the X-Forwarded-Host and X-Forwarded-Prefix headers when building a self link.

This change removes the needs to add the prefix in manually as part of the RS discovery response post-processing, which cleans up the ProcessRs script and route.

https://github.com/SecureApiGateway/SecureApiGateway/issues/874