916: Fixing TokenEndpointTransportCertValidationFilter error responses
When mTLS cert validation is done as part of the token_endpoint AS route, then error responses need to conform to the OAuth 2.0 spec. https://www.rfc-editor.org/rfc/rfc6749#section-5.2

Adding an errorResponse method to TokenEndpointTransportCertValidationFilter which produces errors in the correct format.

TransportCertValidationFilter is used on Open Banking routes, therefore needs to produce an error that conforms to the OB spec.

https://github.com/SecureApiGateway/SecureApiGateway/issues/916