978: Move TestDirectorySigningKeyStore from config.json heap into route heaps
Having TestDirectorySigningKeyStore in config.json means that all environment variables needed to configure this object must be supplied in order for IG to start.

By moving this object into the heap of the routes that use it, then if we omit the config only those routes will be broken.

To disable the test trusted directory fully, it is recommended to remove the routes: 7*-ob-jwkms-*.json from the docker image that is going to be deployed.

https://github.com/SecureApiGateway/SecureApiGateway/issues/978