1066: Consolidating ReverseProxyHandlers which call FR products
We no longer need specific ReverseProxyHandlers for the RS, as the x-ob-url header is no longer required by any APIs. Once this is removed the handlers for FR products look identical.

Consolidating the handlers to: FRReverseProxyHandler and FRReverseProxyHandlerNoCapture

The OBReverseProxyHandler remains unchanged, and should be used when calling out to the OB Directory.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1066