1120: Switching to use the ForgeRockClientHandler in the DCR AddIgAccessTokenForNewRegistrations filter
Previously the TokenRequestHandler was being used, this uses the SettingNewEntity filter to change the client_credentials grant into a password grant and adds the username and password.

client_credentials grant is allowed for this flow, it should be used in preference to password grant whenever possible.

Removing unnecessary scope: fr:idm:* from the gateway access_token request

https://github.com/SecureApiGateway/SecureApiGateway/issues/1120