1121: Moving TokenRequestHandler into IDMClientHandler
TokenRequestHandler is only needed by the IDMClientHandler. Moving it as an inline handler definition so that it cannot be used elsewhere.

Changing the gateway token request from the token endpoint route to use the ForgeRockClientHandler, as it can use client_credentials flow.

Renaming SettingNewEntity => SetIdmPasswordGrantCredentials to make it clearer what the script does.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1121