Adding filter: RouteMetricsFilter which is responsible for gathering and reporting metrics each time a route processes a request.
Installing the filter for the following routes:

- All RS OB routes
- The following AS routes
-- /register
-- /authorize
-- /access_token
Metrics are reported by publishing them to the log file, see LoggerRouteMetricsEventPublisher.

To get the ApiClient data for all of the AS routes has required changes to filters in order to set the ApiClient into the attributes context, the FetchApiClientFilter is used for the RS routes but this does not work with the AS routes as it requires an access_token in the request entity which is not the case. Filters changed:

https://github.com/SecureApiGateway/SecureApiGateway/issues/1151