1179: Setting software_statement correctly in DCR GET response
A bug was introduced in PR: https://github.com/SecureApiGateway/secure-api-gateway-ob-uk/pull/449

As part of refactoring in this PR to make the ApiClient available for the RouteMetricsFilters, a bug was introduced in the DCR GET response. The software_statement field was missing due to the ApiClient in the attributes context having different fields names to the previous implementation.

Fixing ProcessRegistration to call the correct methods on ApiClient to obtain the software_statement.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1179