1104: Extracting JwtReSigner class from ReSignIdTokenFilter
Creating JwtReSigner class which is responsible for re-signing a JWT. It takes a JWT as input, verifies the sig and then produces a JWT using the configured kid and private key (all other header fields and claims remain unchanged).

This logic has been extracted from the ReSignIdTokenFilter, which now depends on an instance of JwtReSigner.

The aim of this refactoring is to enable re-signing of JARM responses, reusing the re-signing logic.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1104