1278: Migrating CreateApiClient.groovy to Java
Created ManageApiClientFilter java impl of CreateApiClient.groovy, the new name captures that the filter does more than just create the ApiClient; it supports the full DCR management API as well.

This filter delegates to ApiClientService and ApiClientOrganisationService to do ApiClient CRUD operations.

ClientIdRequestParameterLocator abstraction has been created so that the client_id can be looked up from the HTTP request. This is necessary as the DCR Management protocol spec (RFC 7592) is experimental and not all Authorisation Servers will implement it in the same fashion, AM for example uses a query param whereas the spec suggests using a path param; implementations for both are provided.

All successful invocations of this filter add the ApiClient to the context so that it can be used by other filters in the chain.

Updated the DCR route to use the new filter.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1278