1227: Gathering java dependencies required by docker using maven
Created new module: secure-api-gateway-ob-uk-docker
This module is responsible for building the docker image, all resources under the top-level docker directory move into this module.

A pom.xml has been added to the module which captures the dependencies which need to be included in the docker image, these include dependencies required by the Groovy scripts and the secure-api-gateway-ig-extensions jar

Removing lib/ and all dependencies checked into source control which were used to build the docker image previously.

Updating Makefile to build the secure-api-gateway-ig-extensions module when creating the docker image, and adding a target which copies the dependencies which need to be included in the image.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1227