1289: Using ResourceOwnerOAuth2ClientFilter to get an IDM access_token
password_grant flow currently needs to be used to get an IDM access_token, updating the IDMClientHandler to use the ResourceOwnerOAuth2ClientFilter to do this.

Removing SetIdmPasswordGrantCredentials.groovy which was a hacky way of doing the password_grant flow by hacking the output of ClientCredentialsOAuth2ClientFilter.

This solution means that the IDMClientHandler will work for all deployments of SAPI-G without the need to have a custom script.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1289