1318: No longer overwrite jwks_uri in as well-known
The OIDC well-known response contains a jwks_uri entry specifying where
the public certs for the TLS and signing certs used by the API provider can be
found. We no longer need to override this as it is now configured in
AM, in the OAuth2Provider's OIDC Advanced config.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/1318