1159: Fixing JWKS Proxy route
After upgrading IG version, we noticed issues with the JWKS proxy route returning errors. This appears to be because of the headers added by nginx are causing issues when the OB infrastructure processes our requests.

Changing the JWKS proxy route to strip off all the incoming headers and to use the OBClientHandler.

Removing the OBReverseProxyHandler as ReverseProxyHandlers should only be used to call protected applications.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1159