1512: Make pass-through route send all traffic to Identity Platform
Rename 99-as-passthrough.json to 999-platform-passthrough.json so that
it is last in the route list. Update route name to reflect this.

The route is now responsible for passing any requests, which have not
already been routed, through to the Identity Platform.

This includes AM endpoints such as the XUI and IDM endpoints to fetch
config used to render the XUI.

Fix name of 98-kube-probe.json route for consistency.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1512