1620: SAPI integration with IG 2024.11 (migration to openig-fapi)
Core SAPI core code is being migrated to core IG - openig-fapi - and this must be reflected in the SAPI core and OBUK codebase (use the migrated IG components as they are available).

Change the pom so that the image is built using:
- IG Docker Image: 2024.11.0
- SAPIG IG Extensions: 4.0.2-SNAPSHOT

We need to use the filters/services provided by v2024.11 of IG
This requires changes to the dynamic client registation route, and of
the process registration scripts.

- Using the config in core as a basis
- Update RCS routes following IG integration
- Update DCR routes following IG integration
- Changed AmService-OBIE to just AmService
- ProcessRegistration.groovy script changes to support integration
- ProcessDetachedSig.groovy script changes to support integration
- Fix config and routes to use JwkSetService not OBJwkSetService (DCR and AS routes)
- Fix config and routes to use TransportCertValidator (AS routes)
- Remove filter for RegistrationRequestJwtSignatureValidationFilter, as this is done within IG and TTD

Notes:
- OBJwkSetService which used to be a part of core SAPIG has moved to IG and been renamed JwkSetService. This PR fixes both config definition of the service and the use of that
service in the register, access_token and par routes.
- OBTransportCertValidator which used to be a part of core SAPIG has moved to IG
and been renamed TransportCertValidator. This PR fixes both config definition of the service and the use of that service in the access_token and par routes.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/1620