1620: Log the route name as a route is called
There is an issue where two requests are being made from a route to the
service being proxied. We want a way to determine if the route is being
called twice. This log entry helps us know which route has been called
when.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/1620